adding after crud

+++ onlinebank_project/src/main/java/com/online/bank/util/GenerateCustomerId.java
@@ -4,10 +4,10 @@ import java.util.Random;
 
 public class GenerateCustomerId {
 
-		public static long generateCustomerId() {
+		public static int generateCustomerId() {
 		Random random = new Random();
-		long nextLong = random.nextLong();
-		return nextLong;
+		int nextInt = random.nextInt();
+		return nextInt;
 	}
 
 }
